February 11, 2025 (Thesis 2 time) Coding with Billy
We got the code generated pa lang. The prettymiditest.py is just for visualization purposes, to check if the notes from the midi files are accurate when read. We weren't able to run the librosa test properly for it (some functionalities were deprecated) so we used the pretty midi library instead (and it works!)