Added -requirements.txt file to be used for virtual environments and uploaded the MTC-ANN-2.0.1 dataset for temporary testing.
In order to use the shared codebase properly, I created a "requirements.txt" file that can be used to install all required dependencies and libraries to run the script, in a local virtual environment.

I also added the dataset in a folder named "dataset" that contains all 360 MIDI files (in .mid format) that are analyzed using the mido Python library.